Allow values besides default root DB usernames
While this isn't helpful when creating your own database, if you were to
make use of something like RDS or other databases as a service, you
usually have the ability to create a root account with the username that
isn't `root` or `postgres`.

With things the way they are, you would be required to always maintain
the above connections to make use of the `bamboo::database` recipe.
This change keeps the original functionality in place, but makes it
possible the change the default through attributes via the following:

    default[:bamboo][:database][:external]       = true
    default[:bamboo][:database][:root_user_name] = 'my_root_username'